% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unternehmenseffekt_korrRegressoren.R
\name{Unternehmeneffekt_korrRegressoren}
\alias{Unternehmeneffekt_korrRegressoren}
\title{Standardfehler bei einem Unternehmenseffekt und korrelierten Regressoren}
\usage{
Unternehmeneffekt_korrRegressoren(
  N = 500,
  T = 10,
  Anzahl_Regressoren = 1,
  Anz_Sim = 100,
  Anteil_u = 0.25,
  Corr_nu = 0.3,
  Corr_mu = 0.3
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{N bestimmt die Anzahl der Jahre im Paneldatensatz}

\item{Anzahl_Regressoren}{Legt die Anzahl der Regressoren fest}

\item{Anz_Sim}{Legt die Anzahl der Simulationen fest. Der Parameter bestimmt wie oft die wahren Standardfehler und die geschätzten für einen Paneldatensatz mit N Unternehmen und T Zeitperioden berechnet werden sollen.}
}
\value{
Für jede Schätzung eines Betas die wahren Standardfehler für die OLS Regression und die Fama-MacBeth Regression
sowie die Schätzung der Standardfehler nach
}
\description{
Die Funktion berechnet für die angegebene Anzahl an Simulationen die wahren Standardabweichungen des OLS und Fama-MacBeth Schätzers und die geschätzten Standardabweichungen nach OLS, Fama-MacBeth, Cluster, Cluster plus Schock und Newey-West und bildet darüber den Durchschnitt.
}
\details{
Da die Anzahl der Regressoren erst in den Argumenten der Funktion bestimmt wird, wird der Nutzer nach dem starten der Funktion aufgefordert, die Korrelationen der Regressoren anzugeben.
}
