% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unternehmenseffekt.R
\name{Unternehmenseffekt}
\alias{Unternehmenseffekt}
\title{Standardfehler bei einem Unternehmenseffekt}
\usage{
Unternehmenseffekt(
  N = 500,
  T = 10,
  sd_X = 1,
  sd_eta = 2,
  Anzahl_Regressoren = 1,
  Anz_Sim = 100,
  Korr_u = 0.25
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{T bestimmt die Anzahl der Perioden im Paneldatensatz}

\item{sd_X}{Standardabweichung der erklärenden Variablen}

\item{sd_eta}{Standardabweichung der Fehlervariablen}

\item{Anzahl_Regressoren}{Anzahl der Regressoren}

\item{Anz_Sim}{Anzahl der Simulationen. Anz_Sim bestimmt, wieviele Paneldatensätze
mit N Unternehmen und T Perioden simuliert werden sollen.}

\item{Korr_u}{Höhe der Korrelation innerhalb der Unternehmenscluster der Fehlervariable
(Anteil der Varianz des Unternehmenseffekts an der gesamten Varianz der Fehlervariable)}
}
\value{
Vektor mit Länge 6. Durchschnittliche Schätzung der Standardfehler nach OLS, Fama-MacBeth, Cluster und Newey-West sowie
die durchschnittlichen wahren Standardfehler der OLS Regression und der Fama-MacBeth Regression
}
\description{
Die Funktion berechnet die geschätzten und wahren Standardfehler bei einem Unternehmenseffekt.
}
\details{
Die Anzahl der Regressoren ist ein Argument der Funktion.
Nach dem Starten der Funtion wird der Nutzer aufgefordert, für jeden Regressor die
Korrelation innerhalb der Unternehmenscluster anzugeben.
}
