% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnternehmensZeiteffekt.R
\name{UnternehmensZeiteffekt}
\alias{UnternehmensZeiteffekt}
\title{Standardfehler bei einem Unternehmens- und Zeiteffekt}
\usage{
UnternehmensZeiteffekt(
  N = 500,
  T = 10,
  Anzahl_Regressoren = 1,
  Anz_Sim = 100,
  Korr_u = 0.25,
  Korr_delta = 0.25
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{T bestimmt die Anzahl der Perioden im Paneldatensatz}

\item{Anzahl_Regressoren}{Anzahl der Regressoren}

\item{Anz_Sim}{Anzahl der Simulationen. Anz_Sim bestimmt, wieviele Paneldatensätze
mit N Unternehmen und T Perioden simuliert werden sollen.}

\item{Korr_u}{Höhe der Korrelation innerhalb der Unternehmenscluster der Fehlervariable
(Anteil der Varianz des Unternehmenseffekts an der gesamten Varianz der Fehlervariable)}

\item{Korr_delta}{Höhe der Korrelation innerhalb der Zeitcluster der Fehlervariable
(Anteil der Varianz des Zeiteffekts an der gesamten Varianz der Fehlervariable)}
}
\value{
Vektor mit Länge 6. Durchschnittliche Schätzung der Standardfehler nach OLS, Fama-MacBeth, Doppelcluster und Newey-West sowie
die durchschnittlichen wahren Standardfehler der OLS Regression und der Fama-MacBeth Regression
}
\description{
Die Funktion berechnet die geschätzten und wahren Standardfehler bei einem Unternehmens- und Zeiteffekts.
}
\details{
Die Anzahl der Regressoren ist ein Argument der Funktion.
Nach dem Starten der Funtion wird der Nutzer aufgefordert, für jeden Regressor die
Korrelation innerhalb der Unternehmens- und Zeitcluster anzugeben.
}
