% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeiteffekt.R
\name{Zeiteffekt}
\alias{Zeiteffekt}
\title{Standardfehler bei einem Zeiteffekt}
\usage{
Zeiteffekt(
  N = 500,
  T = 10,
  Anzahl_Regressoren = 1,
  Anz_Sim = 100,
  Anteil_delta = 0.25
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{T bestimmt die Anzahl der Jahre im Paneldatensatz}

\item{Anzahl_Regressoren}{Legt die Anzahl der Regressoren fest}

\item{Anz_Sim}{Legt die Anzahl der Simulationen fest. Der Parameter bestimmt wie oft die wahren Standardfehler und die geschätzten für einen Paneldatensatz mit N Unternehmen und T Zeitperioden berechnet werden sollen.}

\item{Anteil_delta}{Anteil des Zeiteffekts an der gesamten Varianz der Fehlervariable}
}
\value{
Matrix mit den wahren und geschätzten Standardfehlern
}
\description{
Die Funktion berechnet für die angegebene Anzahl an Simulationen die wahren Standardabweichungen des OLS und Fama-MacBeth Schätzers und die geschätzten Standardabweichungen nach OLS, Fama-MacBeth, Cluster, Cluster plus Schock und Newey-West und bildet darüber den Durchschnitt.
}
\details{
Da die Anzahl der Regressoren erst in den Argumenten der Funktion bestimmt wird, wird der Nutzer nach dem starten der Funktion aufgefordert, die Korrelationen der Regressoren anzugeben.
}
