% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unternehmenseffekt_korrRegressoren.R
\name{Unternehmeneffekt_korrRegressoren}
\alias{Unternehmeneffekt_korrRegressoren}
\title{Standardfehler bei einem Unternehmenseffekt und korrelierten Regressoren}
\usage{
Unternehmeneffekt_korrRegressoren(
  N = 500,
  T = 10,
  Anzahl_Regressoren = 2,
  Anz_Sim = 100,
  Korr_u = 0.25,
  Corr_nu = 0.3,
  Corr_mu = 0.3
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{T bestimmt die Anzahl der Perioden im Paneldatensatz}

\item{Anzahl_Regressoren}{Anzahl der Regressoren}

\item{Anz_Sim}{Anzahl der Simulationen. Anz_Sim bestimmt, wieviele Paneldatensätze
mit N Unternehmen und T Perioden simuliert werden sollen.}

\item{Korr_u}{Höhe der Korrelation innerhalb der Unternehmenscluster der Fehlervariable
(Anteil der Varianz des Unternehmenseffekts an der gesamten Varianz der Fehlervariable)}

\item{Corr_nu}{Höhe der Korrelation zwischen den Komponenten der erklärenden Variablen,
die sich für jedes Unternehmen und jede Periode neu realisieren}

\item{Corr_mu}{Höhe der Korrelation zwischen den Unternehmenseffekten der Regressoren}
}
\value{
Vektor mit Länge 6. Durchschnittliche Schätzung der Standardfehler nach OLS, Fama-MacBeth, Cluster und Newey-West sowie
die durchschnittlichen wahren Standardfehler der OLS Regression und der Fama-MacBeth Regression
}
\description{
Die Funktion berechnet die geschätzten und wahren Standardfehler bei einem Unternehmenseffekt und korrelierten Regressoren.
}
\details{
Die Anzahl der Regressoren ist ein Argument der Funktion.
Nach dem Starten der Funtion wird der Nutzer aufgefordert, für jeden Regressor die
Korrelation innerhalb der Unternehmenscluster anzugeben.
}
