% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unternehmenseffekt_Schock.R
\name{Unternehmenseffekt_Schock}
\alias{Unternehmenseffekt_Schock}
\title{Standardfehler bei Korrelation über die Unternehmenscluster hinaus}
\usage{
Unternehmenseffekt_Schock(
  N = 500,
  T = 10,
  Anz_Sim = 100,
  Korr_mu = 0.25,
  Korr_u = 0.25,
  D_X = 0,
  D_eta = 0,
  L = 10
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{T bestimmt die Anzahl der Perioden im Paneldatensatz}

\item{Anz_Sim}{Anzahl der Simulationen. Anz_Sim bestimmt, wieviele Paneldatensätze
mit N Unternehmen und T Perioden simuliert werden sollen.}

\item{Korr_mu}{Höhe der Korrelation innerhalb der Unternehmenscluster der erklärenden Variable
(Anteil der Varianz des Unternehmenseffekts an der gesamten Varianz der erklärenden Variable)}

\item{Korr_u}{Höhe der Korrelation innerhalb der Unternehmenscluster der Fehlervariable
(Anteil der Varianz des Unternehmenseffekts an der gesamten Varianz der Fehlervariable)}

\item{D_X}{Distanz |i-j| für die, die erklärenden Variablen verschiedener Unternehmen korreliert sind}

\item{D_eta}{Distanz |i-j| für die, die Fehlervariablen verschiedener Unternehmen korreliert sind}

\item{L}{Höhe des L des Discroll-Kraay Schätzers}
}
\value{
Vektor mit Länge 8. Durchschnittliche Schätzung der Standardfehler nach OLS, Fama-MacBeth, Cluster,
Discroll-Kraay, Cluster-Schock und Newey-West sowie
der durchschnittlichen wahren Standardfehler der OLS Regression und der Fama-MacBeth Regression
}
\description{
Die Funktion berechnet die geschätzten und wahren Standardfehler bei erklärenden Variablen und Fehlervariablen,
die über die Unternehmenscluster hinaus korreliert sind.
}
\details{
Da die Anzahl der Regressoren nicht die Fehlschätzung beeinflusst,
simuliert diese Funktion nur Regressionen mit einem Regressor
}
