% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zeiteffekt.R
\name{Zeiteffekt}
\alias{Zeiteffekt}
\title{Standardfehler bei einem Zeiteffekt}
\usage{
Zeiteffekt(
  N = 500,
  T = 10,
  Anzahl_Regressoren = 1,
  Anz_Sim = 100,
  Anteil_delta = 0.25
)
}
\arguments{
\item{N}{N bestimmt die Anzahl der Firmen im Paneldatensatz}

\item{T}{T bestimmt die Anzahl der Perioden im Paneldatensatz}

\item{Anzahl_Regressoren}{Anzahl der Regressoren}

\item{Anz_Sim}{Anzahl der Simulationen. Anz_Sim bestimmt, wieviele Paneldatensätze
mit N Unternehmen und T Perioden simuliert werden sollen.}

\item{Anteil_delta}{Hohe der Korrelation innerhalb der Zeitcluster der Fehlervariable
(Anteil der Varianz des Zeiteffekts an der gesamten Varianz der Fehlervariable)}
}
\value{
Vektor mit Länge 6. Durchschnittliche Schätzung der Standardfehler nach OLS, Fama-MacBeth, Cluster und Newey-West sowie
die durchschnittlichen wahren Standardfehler der OLS Regression und der Fama-MacBeth Regression
}
\description{
Die Funktion berechnet die geschätzten und wahren Standardfehler bei einem Zeiteffekt.
}
\details{
Die Anzahl der Regressoren ist ein Argument der Funktion.
Nach dem Starten der Funtion wird der Nutzer aufgefordert, für jeden Regressor die
Korrelation innerhalb der Zeitcluster anzugeben.
}
